change mother class to extend entity and create texture

+++ core/src/com/medievaltower/entities/monster/Archer.java
@@ -10,7 +10,7 @@ public class Archer extends Monstre {
         super(x, y);
 
         // Créez une texture 1x1 de couleur verte
-        Pixmap pixmap = new Pixmap(this.getWidth(), this.getHeight(), Pixmap.Format.RGBA8888);
+        Pixmap pixmap = new Pixmap((int) this.getWidth(), (int) this.getHeight(), Pixmap.Format.RGBA8888);
         pixmap.setColor(Color.WHITE);
         pixmap.fill();
